[DF] Re-add sanity check to output of variation expression
With the switch to bulk processing the sanity check on the
size of the array returned by the vary expression was
suppressed. This commit re-adds it only checking the
first returned value for each bulk, skipping the rest
of the bulk if there is a problem.

It's a compromise between checking at every entry (expensive)
and never checking (more easily brings execution in UB territory).